Change DiffieHellman

+++ app/src/main/java/com/example/cryptext/data/entity/MyData.kt
@@ -2,6 +2,7 @@ package com.example.cryptext.data.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigInteger
 
 @Entity(tableName = "mydata")
 data class MyData(
@@ -11,5 +12,5 @@ data class MyData(
     val username: String? = null,
     val email: String? = null,
     val password: String? = null,
-    val privateKey: Int,
+    val privateKey: Long,
     )
